Add loop to experiment with resource adding
I set up an xDS server that updates itself from a file,
and then setup this test-runner to continually listen for
new discovery responses.  If I add a new cluster to the
config.yaml, I would expect a new discovery response sent
to this client.  This code sets up an environment to check
my logic on this.