Rewrite Test 1
Setup a test case in a style that allows us to succinctly test LDS and
CDS on the same behaviour.  This required refactoring our runner
functions to work for any service, and to write new step functions to
take advantage of the new way we are passing in parameters.

The new test is designed to fail on 2 of the 6 cases, to assure that a
test can fail but the test suite itself can continue to run.