remove simpleResponse type and functions
We had a parser function that took a discovery response and simplified
it to just the version, nonce, and resource names. It is confusing to be
maintaining two different, but thematically similar, types...especially
when it is easy to check the version and nonce directly from the passed
in discovery response.

I added a resourceNames function that returns just the name values from
a response, then updated any reference to simpleResponse parsing to
instead pull from the direct discovery response and parse names as needed