set timeout to constant and increase
we do a pause in our validation to wait for any lingering responses.
This, combined with a short timeout, can cause tests to fail
erroneously. Increasing the timeout here allows for a clearer error when
there are truly no responses that contain the resources we want, and not
just that the connection ended early.